requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.3
anaconda-project==0.6.0
appdirs==1.4.4
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==2.3.3
astropy==2.0
atomicwrites==1.3.0
attrs==19.3.0
awscli==1.18.211
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bcrypt==3.2.0
beautifulsoup4==4.5.3
bitarray==0.8.1
bkcharts==0.2
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.6
boto==2.47.0
boto3==1.16.51
botocore==1.19.51
Bottleneck==1.2.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.0
cachy==0.3.0
certifi==2019.11.28
cffi==1.10.0
chardet==3.0.2
cleo==0.8.1
click==6.7
clikit==0.6.2
cloudpickle==0.3.1
clyent==1.2.1
colorama==0.3.9
conda==4.3.22
configobj==5.0.6
contextlib2==0.5.5
coverage==5.2.1
crashtest==0.3.1
cryptography==1.9
cycler==0.10.0
Cython==0.26
cytoolz==0.8.2
dask==0.15.1
datashape==0.5.4
decorator==4.1.2
distlib==0.3.1
distributed==1.18.0
docutils==0.13.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.12
firebase-admin==4.0.1
Flask==1.1.1
Flask-Cors==3.0.3
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.7.1
future==0.17.1
gevent==1.2.2
gmpy2==2.0.8
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.12
grpcio==1.28.1
h5py==2.7.0
heapdict==1.0.0
heroku==0.1.4
html5lib==1.1
httplib2==0.17.1
idna==2.5
imageio==2.2.0
imagesize==0.7.1
importlib-metadata==1.6.0
importlib-resources==3.0.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.9
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.11.1
jmespath==0.10.0
jsonschema==2.5.1
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.6
keyring==21.4.0
lazy-object-proxy==0.0.0
llvmlite==0.19.0
locket==0.2.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
more-itertools==8.2.0
mpltools==0.2.0
mpmath==0.19
msgpack==1.0.0
msgpack-python==0.4.8
multipledispatch==0.4.9
mysql-connector-python==8.0.22
mysqlclient==2.0.3
navigator-updater==0.1.0
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.4
nodeenv==1.3.3
nose==1.3.7
notebook==5.0.0
numba==0.34.0
numexpr==2.6.2
numpy==1.13.1
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.5.0a2
packaging==20.4
pandas==0.20.3
pandocfilters==1.4.1
partd==0.3.8
pastel==0.2.1
path.py==8.2.1
pathlib2==2.3.0
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.8.0
pickleshare==0.7.3
Pillow==4.2.1
pipenv==2020.6.2
pkginfo==1.6.1
pluggy==0.13.1
ply==3.10
poetry==1.1.4
poetry-core==1.0.0
prompt-toolkit==1.0.14
protobuf==3.11.3
psutil==5.2.1
ptyprocess==0.5.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.1
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pydocstyle==4.0.1
pyflakes==1.5.0
Pygments==2.2.0
pygpu==0.6.9
PyJWT==1.4.2
pylev==1.3.0
pylint==2.4.1
pyodbc==4.0.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
pytest==5.2.0
python-dateutil==2.8.1
Python-dev==2.0.0.dev0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==5.3.1
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.22.0
requests-toolbelt==0.9.1
rope==0.10.5
rsa==4.0
ruamel-yaml==0.11.14
s3transfer==0.3.3
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.1
seaborn==0.8
selenium==3.141.0
sh==1.12.14
shellingham==1.3.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==1.2.1
sortedcollections==0.4.2
sortedcontainers==1.5.3
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.0
SQLAlchemy==1.3.22
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.2.0
terminado==0.6
testpath==0.3
Theano==0.9.0
tomlkit==0.7.0
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
twilio==6.38.0
typed-ast==1.4.1
typing==3.6.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.0.28
virtualenv-clone==0.5.4
wcwidth==0.1.7
webencodings==0.5
Werkzeug==1.0.1
widgetsnbextension==2.0.0
wrapt==1.11.2
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.11.4
xlwt==1.2.0
zict==0.1.2
zipp==3.1.0
